Initial commit to pre-processing branch
Changes:
Created a new file, brought in data, did some preliminary visualization and cleaning. Some mild feature engineering

Upcoming:
More feature engineering to come in the coming days, as well as generalization and robustness development of the methods and procedure into functions to be used and called easily in the implementation steps.