Creating functions and implmenting modularity
Created a new file.
Utilizes functions to make stock dealing easy and modular. Only specify the filepath, the type of stock (stock or ETF), and the abbreviation of the stock. The nested functions handle the rest.

Easily modular, and relatively easy to follow. Adapted from the work of my own, the work from Patrick in his initial commit, and the some of the work from Rajnish.