Changed file folder organization and structure
Reorganized the file folder, and moved things into proper subfolders. Copies of files are still present  in the dev_files folder so that not work is lost in the case that we need to use something which is not included in the main script. The most up-to-date preprocessing is in preprocessing>usable_functions