lstm updates
I have now fixed the issues with what shape all the vectors should be and this helped it run better. I also now allow an output layer that allows different number of hiddens to be in the short and long memory and then the output layer projects them in the desired number of outputs. I also fixed the model because it wasn't using all the input data but instead was stopping one short in the sequence. I am now testing to see the overall improvement because before the model wasn't decreasing it's loss very well at all.