lstm testing
I have tested the model against the pytorch model and they both perform about the same... which is to say they both suck but hey at least it's not just us. I also think we should drop the Y = H*w + b_o layer as it doesn't seem to help the predictions very much. So this means that the number of hiddens should be the output number