updated the history traces
the x_hist was broken for 2d inputs so I have fixed it now to be able to taken input slices. This makes x_hist dimensions (#features, #samples). This is directly a copy of the input tensor but is a numpy array. We may not need this but it is there in case we do.